move src/utils to spinel/spinel
