+++ BUILD.gn
@@ -162,9 +162,9 @@ static_library("lib-ot-core") {
     "src/core/utils/parse_cmdline.cpp",
     "src/core/utils/slaac_address.cpp",
     "src/ncp/hdlc.cpp",
-    "src/utils/spinel.c",
-    "src/utils/spinel_decoder.cpp",
-    "src/utils/spinel_encoder.cpp",
+    "src/spinel/spinel.c",
+    "src/spinel/spinel_decoder.cpp",
+    "src/spinel/spinel_encoder.cpp",
     "third_party/mbedtls/repo/library/md.c",
     "third_party/mbedtls/repo/library/md_wrap.c",
     "third_party/mbedtls/repo/library/memory_buffer_alloc.c",
