+++ src/CMakeLists.txt
@@ -29,4 +29,4 @@
 add_subdirectory(cli)
 add_subdirectory(core)
 add_subdirectory(ncp)
-add_subdirectory(utils)
+add_subdirectory(spinel)
