+++ src/spinel/spinel_encoder.hpp
