+++ src/spinel/spinel_encrypter.hpp
