+++ tests/unit/test_spinel_decoder.cpp
@@ -28,7 +28,7 @@
 
 #include "common/code_utils.hpp"
 #include "common/instance.hpp"
-#include "utils/spinel_decoder.hpp"
+#include "spinel/spinel_decoder.hpp"
 
 #include "test_util.hpp"
